requirements.txt
opencv-python
tensorflow==2.1
tensorflow-gpu==2.1
numpy
imutils==0.5.3
PyYaml
playsound
mutagen
